added balance and list withdrawals handlers

+++ internal/model/user.go
@@ -8,6 +8,6 @@ type User struct {
 	ID        uint      `db:"id"`
 	Login     string    `db:"login" validate:"required"`
 	Password  string    `db:"password" validate:"required"`
-	Balance   int       `db:"balance"`
+	Balance   float64   `db:"balance"`
 	CreatedAt time.Time `db:"created_at"`
 }
